+++ app/handlers/newflow/oauth_callback.rb
@@ -42,13 +42,15 @@ module Newflow
       elsif mismatched_authentication?
         fatal_error(code: :mismatched_authentication)
       elsif (outputs.authentication = Authentication.find_by(provider: @oauth_provider, uid: @oauth_uid))
-        # User found with the given authentication.
-        # We will log them in.
+        # User found with the given authentication. We will log them in.
+        BRI_marketing(outputs.authentication.user)
       elsif (existing_user = user_most_recently_used(users_matching_oauth_data))
         # No user found with the given authentication, but a user *was* found with the given email address.
         # We will add the authentication to their existing account and then log them in.
         outputs.authentication = Authentication.find_or_initialize_by(provider: @oauth_provider, uid: @oauth_uid)
         run(TransferAuthentications, outputs.authentication, existing_user)
+
+        BRI_marketing(existing_user)
       elsif (old_flow_user = find_old_flow_user(provider: @oauth_provider, uid: @oauth_uid))
         # create a corresponding new flow Authentication
         outputs.authentication = create_newflow_auth_for_user(
@@ -56,6 +58,7 @@ module Newflow
           provider: @oauth_provider,
           uid: @oauth_uid
         )
+        BRI_marketing(old_flow_user)
       elsif user_came_from&.to_sym == LOGIN_FORM_IS_ORIGIN
         # The user is trying to sign up but they came from the login form, so redirect them to the sign up form
         fatal_error(code: :should_redirect_to_signup)
@@ -69,6 +72,14 @@ module Newflow
     end
 
     private ###########################
+
+    def BRI_marketing(user)
+    # If user is BRI (Bill of Rights Institute) book adopter, we want to track that and do marketing
+      if options[:is_BRI_book]
+        user.update!(is_b_r_i_user: true)
+        UpdateSalesforceLead.perform_later(user: user)
+      end
+    end
 
     # users can only have one login per social provider, so if user is trying to log in with
     # the same provider but it has a different uid, then they might've gotten the social account hacked,
