+++ spec/routines/newflow/student_signup/activate_student_spec.rb
@@ -21,7 +21,7 @@ module Newflow
 
         it 'does NOT sign up user for the newsletter when NOT checked' do
           user.update(receive_newsletter: false)
-          expect_any_instance_of(PushSalesforceLead).to receive(:exec).with(hash_including(newsletter: false))
+          expect_any_instance_of(Newflow::CreateSalesforceLead).to receive(:exec)
           described_class.call(user)
         end
 
@@ -29,9 +29,8 @@ module Newflow
           source_app = FactoryBot.create(:doorkeeper_application)
           user.update(source_application_id: source_app.id)
 
-          expect_any_instance_of(PushSalesforceLead).to(
-            receive(:exec).with(hash_including(source_application: source_app))
-          )
+          expect_any_instance_of(Newflow::CreateSalesforceLead).to receive(:exec)
+
           described_class.call(user)
         end
       end
